refactor(backend): Split training.rs into smaller modules
Moved constants, types, compute logic, and plate calculators into separate files under the `training` module. This improves maintainability and reduces file complexity while keeping the public API identical via re-exports.
